fastx-toolkit 0.0.14: add "opt" feature support
(Lab7 ticket: BIOBUILDS-290)

+++ fastx-toolkit/0.0.14/build.sh
@@ -1,25 +1,44 @@
 #!/bin/bash
 
-[ "$BB_ARCH_FLAGS" == "<UNDEFINED>" ] && BB_ARCH_FLAGS=
-[ "$BB_OPT_FLAGS" == "<UNDEFINED>" ] && BB_OPT_FLAGS=
-[ "$BB_MAKE_JOBS" == "<UNDEFINED>" ] && BB_MAKE_JOBS=1
-CFLAGS="${CFLAGS} ${BB_ARCH_FLAGS} ${BB_OPT_FLAGS}"
-CXXFLAGS="${CXXFLAGS} ${BB_ARCH_FLAGS} ${BB_OPT_FLAGS}"
+# Pull in the common BioBuilds build flags
+BUILD_ENV="${PREFIX}/share/biobuilds-build/build.env"
+if [[ ! -f "${BUILD_ENV}" ]]; then
+    echo "FATAL: Could not find build environment configuration script!" >&2
+    exit 1
+fi
+source "${BUILD_ENV}" -v
 
 cp -f "${PREFIX}/share/autoconf/config.guess" config/config.guess
 cp -f "${PREFIX}/share/autoconf/config.sub" config/config.sub
 
-if [ `uname -s` == 'Darwin' ]; then
+if [ "${BUILD_OS}" == 'Darwin' ]; then
     # Needed to ensure libgcc is found when running "make check"
     export DYLD_FALLBACK_LIBRARY_PATH="${PREFIX}/lib"
 fi
 
-env CFLAGS="${CFLAGS}" CXXFLAGS="${CXXFLAGS}" LDFLAGS="${LDFLAGS}" \
+if [[ ${CC_IS_GNU} -eq 1 ]]; then
+    GCC_VER=`"${CC}" -dumpversion`
+    GCC_MAJOR_VER=$(cut -d. -f1 <<<"$GCC_VER")
+    GCC_MINOR_VER=$(cut -d. -f2 <<<"$GCC_VER")
+
+    # Use older, pre-C++11-compliant ABI when building with newer versions of
+    # g++ to maintain compatibility with libgtextutils; omitting this will
+    # cause std::string-related undefined symbol errors. For details, see:
+    #   - https://gcc.gnu.org/onlinedocs/libstdc++/manual/using_dual_abi.html
+    #   - https://gcc.gnu.org/onlinedocs/libstdc++/manual/using_macros.html
+    if [[ $GCC_MAJOR_VER -ge 5 ]]; then
+        CXXFLAGS="${CXXFLAGS} -D_GLIBCXX_USE_CXX11_ABI=0"
+    fi
+fi
+
+env CC="${CC}" CFLAGS="${CFLAGS}" \
+    CXX="${CXX}" CXXFLAGS="${CXXFLAGS}" \
+    LDFLAGS="${LDFLAGS}" \
     ./configure --prefix="${PREFIX}" \
     --enable-wall --disable-debug \
     --with-pic --enable-shared --enable-static \
     2>&1 | tee configure.log
-make -j${BB_MAKE_JOBS}
+make -j${MAKE_JOBS} V=1 2>&1 | tee build.log
 make check
 
 make install
